add first failing test for tdd approach. change parameter type from double to int

+++ src/main/kotlin/jalil/kotlintdd/ChargeDisplay.kt
@@ -1,7 +1,7 @@
 package jalil.kotlintdd
 
 class ChargeDisplay {
-    fun displayCharge(currentCharge: Double) {
+    fun displayCharge(currentCharge: Int) {
         // do things
     }
 }
